refactor move selectRoomProfiles to rooms
