+++ src/selectors/rooms.ts
@@ -4,6 +4,7 @@ import {AppState} from '../store';
 import {initialState} from '../slices/rooms';
 
 const roomsDomain = (state: AppState) => state.rooms || initialState;
+const profilesDomain = (state: AppState) => state.profiles || initialState;
 
 export const getAllRooms = createSelector(
   [roomsDomain],
@@ -17,4 +18,12 @@ export const getSelectedRoom = createSelector([roomsDomain], (roomsState) =>
 export const getUserIds = createSelector(
   [roomsDomain],
   (roomState) => roomState.userIds,
+);
+
+export const selectRoomProfiles = createSelector(
+  [profilesDomain, roomsDomain],
+  (profileState, roomState) =>
+    Array.from(profileState.profiles.values()).filter((profile) =>
+      roomState.userIds.find((id) => id === profile.id),
+    ),
 );
